LRIS-31779 capitalize class names & move vendor js to its own folder
