+++ src/js/components/LoadingIndicator.js
