+++ src/js/index.js
@@ -1,7 +1,7 @@
 import '../css/main.scss';
 
-import './components/documentation-side-nav.js'
-import './zendesk-script.js';
+import './components/DocSideNav.js'
+import './vendor/zendesk-script.js';
 
 export {default as render} from './helpers/preact-renderer.js';
-export {default as DocSideNav} from './components/documentation-side-nav.js';
+export {default as DocSideNav} from './components/DocSideNav.js';
